JT-113：JT-123 ：core-ng-api

+++ db-service-db-migration/src/main/resources/db/migration/R__productions.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS `productions` (
   `id`           VARCHAR(50)                 NOT NULL,
   `status`       VARCHAR(20)                 NOT NULL,
-  `price`        DOUBLE                      NOT NULL,
+  `price`        DECIMAL                     NOT NULL,
   `created_time` TIMESTAMP                   NOT NULL,
   `updated_time` TIMESTAMP,
   `created_by`   VARCHAR(50)                 NOT NULL,
